requirements.txt
Django>=5.0
pandas
PyVCF3
matplotlib
seaborn
numpy
